Decks table view controller refactoring.

+++ flashcards/DecksTableViewController.swift
@@ -36,29 +36,32 @@ class DecksTableViewController: UIViewController, UITableViewDataSource, UITable
         
         // Segue to the DeckViewController
         if segue.identifier == "toDeck" {
-            if let cell = sender as? UITableViewCell, let indexPath = decksTable.indexPath(for: cell) {
-                let deckViewController = segue.destination as! DeckViewController
+            if  let cell = sender as? UITableViewCell, let indexPath = decksTable.indexPath(for: cell),
+                let deckViewController = segue.destination as? DeckViewController {
                 let deck = decks[indexPath.row]
                 deckViewController.deck = deck
             }
         }
         if segue.identifier == "toAppUsage" {
-            let appUsageViewController = segue.destination as! AppUsageViewController
- 
+            guard let appUsageViewController = segue.destination as? AppUsageViewController else { return }
             appUsageViewController.decks = decks
         }
     }
+    
     // TableView methods
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return decks.count
     }
  
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "deckCell") as! DecksTableViewCell
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: "deckCell") as? DecksTableViewCell else {
+            return UITableViewCell()
+        }
+        
         let deck = decks[indexPath.row]
         
         cell.accessoryType = UITableViewCellAccessoryType.disclosureIndicator
-        cell.title.text = deck.name!
+        cell.title.text = deck.name
         cell.badge.text = getBadgeNumberString(deck: deck)
         
         return cell
@@ -72,12 +75,17 @@ class DecksTableViewController: UIViewController, UITableViewDataSource, UITable
     
     // Gets the number of reviews to do for a deck
     func getBadgeNumberString(deck: Deck) -> String {
-        let calendar = Calendar.current
-        let tomorrowDate = calendar.date(byAdding: .day, value: 1, to: Date())
-        let tomorrowTimestamp = calendar.startOfDay(for: tomorrowDate!).timeIntervalSince1970
+        guard let tomorrowDate = Calendar.current.date(byAdding: .day, value: 1, to: Date()) else {
+            return "0"
+        }
+        
+        let tomorrowTimestamp = Calendar.current.startOfDay(for: tomorrowDate).timeIntervalSince1970
         let datePredicate = NSPredicate(format: "nextShown < %f", tomorrowTimestamp)
         let filteredDeck = deck.cards?.filtered(using: datePredicate) as NSSet?
-        let cardsSet = filteredDeck! as! Set<Card>
+        guard let cardsSet = filteredDeck as? Set<Card> else {
+            return "0"
+        }
+        
         let count = cardsSet.count
         
         return "\(count)"
@@ -90,16 +98,18 @@ class DecksTableViewController: UIViewController, UITableViewDataSource, UITable
             
             let deleteAction = UIAlertAction(title: "Delete", style: .destructive, handler: { (action) in
                 
-                let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-                let AppDelegate = UIApplication.shared.delegate as! AppDelegate
-                let deck = self.decks[indexPath.row]
+                if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
+                    let context = appDelegate.persistentContainer.viewContext
+                    let deck = self.decks[indexPath.row]
+                    
+                    context.delete(deck)
+                    appDelegate.saveContext()
                 
-                context.delete(deck)
-                AppDelegate.saveContext()
-            
-                self.fetchDecks()
-                self.decksTable.reloadData()
+                    self.fetchDecks()
+                    self.decksTable.reloadData()
+                }
             })
+            
             alertController.addAction(deleteAction)
             
             let cancelAction = UIAlertAction(title: "Cancel", style: .default, handler: { (action) in
@@ -116,7 +126,9 @@ class DecksTableViewController: UIViewController, UITableViewDataSource, UITable
     
     // Fetch decks from Core data entity Deck
     func fetchDecks() {
-        let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
+        guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return }
+        
+        let context = appDelegate.persistentContainer.viewContext
         
         do {
             decks = try context.fetch(Deck.fetchRequest())
@@ -129,7 +141,6 @@ class DecksTableViewController: UIViewController, UITableViewDataSource, UITable
             
             self.present(alert, animated: true, completion: nil)
         }
-        
     }
